+++ R/fsd_fit.R
@@ -1,4 +1,4 @@
-#' Fit the floc model.
+#' Fit the fsd model.
 #'
 #' Fit the Savin & Doyle's localization error model with fBM process (See **Details**).
 #'
@@ -6,13 +6,13 @@
 #' @template args-dt
 #' @template args-Tz
 #' @template args-vcov
-#' @return A vector of estimated parameters on transformed scale (See [floc_model()]). If `vcov`, a list with components:
+#' @return A vector of estimated parameters on transformed scale (See [fsd_model()]). If `vcov`, a list with components:
 #' \describe{
 #' \item{coef}{A vector of estimated parameters on transformed scale.}
 #' \item{vcov}{A matrix of estimated covariance of parameters on transformed scale.}
 #' }
 #'
-#' @details The Savin & Doyle's localization error model with fBM process (floc model) has the following form:
+#' @details The Savin & Doyle's localization error model with fBM process (fsd model) has the following form:
 #' \deqn{
 #' Y_n = 1/\tau \int_0^\tau X_{n \Delta t +s} ds + \sigma e_{n}
 #' }{
@@ -27,16 +27,16 @@
 #' When put `sigma2` = 0, this model becomes fractional dynamic error model.
 #' When put `(tau, sigma2)` = 0, this model becomes fractional Brownian model.
 #'
-#' Optimization is done by [optimize()]. It works better when parameters are re-parametrized into unrestricted form (See [floc_model()]).
+#' Optimization is done by [optimize()]. It works better when parameters are re-parametrized into unrestricted form (See [fsd_model()]).
 #'
 #' @references Savin, Thierry, and Patrick S. Doyle. "Static and dynamic errors in particle tracking microrheology." Biophysical journal 88.1 (2005): 623-638.
 #'
 #' @example examples/fit_setup.R
-#' @example examples/floc_fit.R
+#' @example examples/fsd_fit.R
 #'
 #' @export
-floc_fit <- function(dX, dt, Tz, vcov = TRUE) {
-  model <- floc_model()
+fsd_fit <- function(dX, dt, Tz, vcov = TRUE) {
+  model <- fsd_model()
   ans <- csi_fit(model, dX, dt, Tz, vcov)
   ans
 }