+++ examples/floc_acf.R
@@ -1,4 +1,0 @@
-tau <- 0.1
-sigma2 <- 0.001
-# ACF of floc model
-floc_acf(alpha, tau, sigma2, dt, N)