+++ examples/floc_fit.R
@@ -1,2 +1,0 @@
-# Fit the fLOC model
-floc_fit(dX, dt, Tz, vcov = TRUE)