+++ examples/fsd_acf.R
@@ -1,0 +1,4 @@
+tau <- 0.1
+sigma2 <- 0.001
+# ACF of fsd model
+fsd_acf(alpha, tau, sigma2, dt, N)