+++ examples/fsd_fit.R
@@ -1,0 +1,2 @@
+# Fit the fsd model
+fsd_fit(dX, dt, Tz, vcov = TRUE)