+++ tests/testthat/test-csi_resid.R
@@ -9,15 +9,15 @@ csi_fwd <- function(Z, dt, mu, acf, Sigma) {
   t(t(dX) + mu * dt)
 }
 
-racf <- function(N, dt, type = c("fbm", "floc", "farma")) {
+racf <- function(N, dt, type = c("fbm", "fsd", "farma")) {
   alpha <- runif(1)
   type <- match.arg(type)
   if(type == "fbm") {
     acf1 <- fbm_acf(alpha, dt, N)
-  } else if(type == "floc") {
+  } else if(type == "fsd") {
     tau <- runif(1)
     sig <- rexp(1)
-    acf1 <- floc_acf(alpha, tau, sig^2, dt, N)
+    acf1 <- fsd_acf(alpha, tau, sig^2, dt, N)
   } else if(type == "farma") {
     phi <- rho <- runif(1,-1,1)
     acf1 <- farma_acf(alpha, phi, rho, dt, N)
@@ -34,7 +34,7 @@ test_that("Z == csi_resid(dX = csi_fwd(Z))", {
     dt <- runif(1)
     N <- sample(1000:2000,1)
     Z1 <- matrix(rnorm(N*nd),N,nd)
-    acf1 <- racf(N, dt, "floc")
+    acf1 <- racf(N, dt, "fsd")
     dX <- csi_fwd(Z1, dt, mu, acf1, Sigma)
     Z2 <- csi_resid(dX, dt, mu, acf1, Sigma)
     expect_equal(max(abs(Z1-Z2)), 0)